+++ Assets/js/App.vue
@@ -1,7 +1,10 @@
 <template>
   <div :class="{'nav-open': $sidebar.showSidebar}">
     <notifications></notifications>
-    <router-view></router-view>
+    <dashboard>
+      <slot></slot>
+    </dashboard>
+    <!--router-view></router-view-->
   </div>
 </template>
