+++ Assets/js/plugins/debug-mode.js
@@ -1,0 +1,10 @@
+const DebugMode = {
+  install (Vue) {
+    if (process.env.NODE_ENV !== 'production') {
+      Vue.config.debug = true
+      Vue.config.devtools = true
+    }
+  }
+}
+
+export default DebugMode
