+++ Assets/js/plugins/sidebar.js
@@ -1,0 +1,3 @@
+import SideBar from '../components/UIComponents/SidebarPlugin'
+
+export default SideBar
