+++ TaxApp/dashboard/templates/dashboard/base.html
@@ -5,6 +5,7 @@
     <meta charset="utf-8">
     <link rel="shortcut icon" type="image/vnd.microsoft.icon" href="{% static 'images/favicon.ico' %}">
     <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
+    <meta name="csrftoken" content="">
 
     <title>
       Joint Tax Boars {% block pageTitle %}{% endblock pageTitle %}
